requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
Babel==2.12.1
backoff==2.2.1
bibtexparser==1.4.0
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi>=2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
colorlog==6.7.0
confection==0.1.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
dgl==1.1.2
dglgo==0.0.2
dill==0.3.7
docutils==0.20.1
emoji==2.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
filelock==3.12.3
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
func-timeout==4.3.5
huggingface-hub==0.16.4
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
lemminflect==0.2.3
Levenshtein==0.21.1
line-profiler==4.1.1
line-profiling @ git+https://github.com/hitachi-nlp/line-profiling.git@edc470335047d6ebf09db968dcf1bf43397538a2
lit==16.0.6
littleutils==0.2.2
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy>=1.22.0
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ogb==1.3.6
openai==0.27.9
outdated==0.2.2
packaging==23.1
pandas==2.1.0
pathy==0.10.2
Pillow==10.0.0
ply==3.11
preshed==3.0.8
protobuf==4.24.3
psutil==5.9.5
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pylev==1.4.0
pyparsing==3.0.9
pyswip==0.2.10
python-constraint==1.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==3.3.0
rdkit-pypi==2022.9.5
-e git+https://github.com/yuchenlin/rebiber.git@a4e024042f56bf665b9c9f53974b5652c261f79e#egg=rebiber
regex==2023.8.8
requests>=2.28.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
scitools-pyke==1.1.1
six==1.16.0
sklearn==0.0.post7
smart-open==6.4.0
snowballstemmer==2.2.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
srsly==2.4.7
stanza==1.5.1
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
timeout-decorator==0.5.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tqdm>=4.64.0
transformers==4.33.1
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.7
yarl==1.9.2
z3-solver==4.12.2.0
zipp==3.16.2
argparse>=1.4.0
